requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
numpy>=1.26.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
flask>=3.0.0
flask-socketio>=5.3.0
pyttsx3>=2.90
elevenlabs>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
ddgs>=0.1.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
uuid>=1.30
